% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_Visualize_contact_tracing_data.R
\name{Visualize_contact_tracing_data}
\alias{Visualize_contact_tracing_data}
\title{Visualize contact tracing data}
\usage{
Visualize_contact_tracing_data()
}
\value{
R Shiny application.
}
\description{
Run an application that visualizes contact tracing data that can be used to
estimate incubation and latency time using the doublIn package.
}
\details{
We have used this app to visualize contact tracing data that was collected
using a tailor-made contact tracing form. For details and source code, have
a look at the Github page of Manh Nguyen Duc.
}
\examples{
# To run the app:
if(interactive()){Visualize_contact_tracing_data()}

}
\references{
Manh Nguyen Duc (2024) Contact tracing form using KoboToolbox
https://github.com/manhnguy/Contact-Tracing-for-Respiratory-Transmitted-Diseases
}
\author{
Vera Arntzen, \email{v.h.arntzen@math.leidenuniv.nl}
}
\keyword{survival}
